Return serverId for backups in Rackspace driver
Backups on the Rackspace Cloud are stored as images, so they can be seen
in response to list_images. However, as they are generically named, it's
currently impossible to tell of which Node they're a backup.

The Rackspace Cloud API returns the serverId of the origin Node. Add
this to the extra dict of the image.

git-svn-id: https://svn.apache.org/repos/asf/incubator/libcloud/trunk@886643 13f79535-47bb-0310-9956-ffa450edef68