Make powered off Linodes appear as TERMINATED
The proper state for powered off is actually TERMINATED, since a
large number of nodes at CK were stuck in REBOOTING (but were
really powered off for good).

Note that this change might make TERMINATED appear briefly during
a reboot, as a rebooting Linode is powered off for a brief window.

git-svn-id: https://svn.apache.org/repos/asf/incubator/libcloud/trunk@984148 13f79535-47bb-0310-9956-ffa450edef68