Import Load Balancers support.
At the moment drivers for two providers are available:
Rackspace and GoGrid. At the moment only basic features
provided by load balancing service are supported, but
they've been tested and should work without problems.

However, there are a lot of work to be done here still,
like adding locations support in a similar fashion as it's
done for compute part, adding blocking mode and refining
implementation in general.

Follow discussion on the maillist to be up-to-date with
these efforts:

http://mail-archives.apache.org/mod_mbox/incubator-libcloud/201104.mbox/<20110404221031.GA26577@reemsky>


git-svn-id: https://svn.apache.org/repos/asf/incubator/libcloud/trunk@1095180 13f79535-47bb-0310-9956-ffa450edef68