Fix chicken-and-egg problem.
I noticed this problem frequently: a Connection's request() instantiates
and response, then sets the responses connection attr to itself;
however, many of the Responses have __init__s that call their parse_body
or parse_error which refers to its own connection object to get details
for error output. Fail. Have to pass the connection to the constructor
to be available early enough.