Let httplib set header.
HTTP spec states that the Host header must contain the port number if
it's non-standard (see
http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.23). The
incorrect override that was present here was breaking the managerment
urls in OpenStack auth responses. httplib does the right thing if we let
it.

Also, the "connection =" removal was because it's shadowed by a
following assignment.