Remove Auth 1.1 support and allow tenant_id.
Auth 1.1 is not necessary. Auth 1.1 servers accept 1.0 authentication,
which gives just as much data (i.e. the same management API URLs).
Note: I'll fix the tests later.

The Connections now also allow a tenant_id (which appears to be the new
name for project_id, so I'll go with it) to be passed in and honored
(this is necessary for Alpha customers).