Issue LIBCLOUD-402: Support OS-DCF:diskConfig
Rackspace is implementing a change to server creation where disk
partitioning will not automatically fill the entire disk, leaving the
option to the user to expand the disk or not. When creating nodes, you
have the ability to specify if you want a MANUAL disk config, or if you
want an AUTO disk config. The former results in quicker server creation,
while the latter results in the entire disk being used.

This change is exposed via an extension to the OpenStack compute APIs,
which Rackspace's Cloud Servers are built on top of.
"OS-DCF:diskConfig" is an attribute to be set in message bodies POST'ed
to /servers, and is exposed as the "ex_disk_config" parameter to
OpenStack_1_1_NodeDriver.create_node.

Signed-off-by: Tomaz Muraus <tomaz@apache.org>