Fix F401 errors from flake8
flake8 emits warnings for names that are imported but not used. In most
cases, these either look like copy/pasted import blocks, or names left
in from debugging or after refactoring. Common unused names are
the sys or time modules, and base classes. Some modules were doing an
import dance to get simplejson or standard json, but those modules no
longer parse json directly, likely as the result of refactoring it to a
lower level.

The libcloud.utils.py3 module contains a lot of these unused imports,
but that's intentional. Because of that, it includes a "flake8: noqa"
comment near the top of the file to tell flake8 to ignore it.