ec2 driver: Make cidr_ips argument mandatory
If the cidr_ips parameter in this function is None, and therefore the
IpPermissions.1.IpRanges.1.CidrIp is not present in the request to the EC2 API,
the request is not performed. (The response does not indicate an error, but the
security rule is not created).

This behaviour is not documented in
http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AuthorizeSecurityGroupIngress.html,
but it is reproducable.

The following request didn't work:

    Action=AuthorizeSecurityGroupIngress
    GroupId=sg-a4b44ece
    IpPermissions.1.FromPort=22
    IpPermissions.1.IpProtocol=tcp
    IpPermissions.1.ToPort=22
    AWSAccessKeyId=...
    Signature=...
    SignatureMethod=HmacSHA256
    SignatureVersion=2
    Timestamp=2014-04-16T07%3A43%3A21Z
    Version=2013-10-15

While this one did (the only difference is adding CidrIp):

    Action=AuthorizeSecurityGroupIngress
    GroupId=sg-a4b44ece
    IpPermissions.1.FromPort=22
    IpPermissions.1.IpProtocol=tcp
    IpPermissions.1.ToPort=22
    IpPermissions.1.IpRanges.1.CidrIp=0.0.0.0/0
    AWSAccessKeyId=...
    Signature=...
    SignatureMethod=HmacSHA256
    SignatureVersion=2
    Timestamp=2014-04-16T07%3A43%3A21Z
    Version=2013-10-15