GCE HTTP-LB support part 3: rebase & consistency
This comment rebases HTTP-LB back up to trunk and renames new methods
to be consistent with the convention of targetinstances.

Also of note is that GCE's ex_create_forwarding_rule has backwards
compatible parameter changes to rename 'targetpool' to the more generic
'target' now that it's used by targetinstance & targethttpproxy.