Reset file's current position in read_in_chunks()
If a cloud storage driver's parameter  supports_chunked_encoding is False,  the function read_in_chunks will be called twice during the upload_object_via_stream progress. But in the second time, the current position of the file to be uploaded is the end, so the StopIteration exception will be raised and upload a empty file.