Cache locations to improve list_images performance
This change caches the locations in a private attribute to avoid
repeated calls for the same result set. The number of calls made is
directly proportional to the number of images returned. For a 223
result set, this can reach as high as 99.990 seconds. Caching the
locations reduces it to around 2.001 seconds. The risk of stale location
data during execution is negligible.