[google compute] fix pricing script and test; fix build
Revert conditional import of simplejson to unconditional: the code as
implemented only works with simplejson and not the standard json library because
it relies on the `item_sort_key` functionality which only simplejson provides
and json does not.

Also fixes indentation which breaks the build since the linter is part of the
tests.

Fixes the issues seen here:
* https://travis-ci.org/apache/libcloud/builds/90117628
* https://travis-ci.org/apache/libcloud/jobs/90117643

Closes #630

Signed-off-by: Eric Johnson <erjohnso@google.com>