EC2: Sort pricing by (non-)numeric part
contrib/scrape-ec2-prices.py changes the ordering of data/pricing.json,
as it sorts the dictionary keys by string, which puts '10' before '2'.

Add a function to split the dictionary keys into the numeric, alphabetic
and other part and compare the first as integer, the middle by meaning,
and the later as string.