Fix an edge case bug when consuming data from a channel inside the Paramiko SSH client.
In case consumed data contained multi byte UTF-8 character and this character
was split over two chunks, an exception would be thrown since we tried to
decode received data after receiving every chunk instead of the end.

This bug was originally caught by Lakshmi Kannan and fix is partially based
on the changes inside StackStorm/st2 which are Apache 2.0 licensed.