Fix openstack v3 authentication.
This change also allows user to provide a custom value for the OpenStack
"domain" which defaults to "Default".

With the code for the OpenStack Identity API v3, two new parameters were added:
domain_name and token_scope but it was impossible to define them to a value
other than their respective default.

Closes #744