Added a mock virConnect class
in this I took all the functions defined on a libvirt.virConnect
class and added them to the class as a stub function.

the stub function will take any number of arguments and return 0

making creation and API calls super fast ;) for testing purposes
and we can add correct Mocking behavior later.