auroradns: Implement iterator for zones and records
By implementing iterate_zones() and iterate_records() you can
iterate over zones and records returned by the API.

The list_zones() and list_records() methods simply talk to the
iteration functions and create a list of it and return that.

The test cases should still match since they call the list functions
and those call the iterators on their turn.

The test case test_list_zones() however has been expanded to iterate
over the result.

The test_list_records case already iterated over the results, so that
verifies that the result is iteratable.

Closes #829