Add Outscale API constants working with TINA OS
Update version and namespace matching the current TINA OS API (AWS EC2
compatible)

This patch handle two different versions between the API version used to issue
requests and the API version crafted in the namespace. It is a bug in the
server side currently. The bugfix is work in progress. Meanwhile we need those
bits in place to keep the new APIs working.

Signed-off-by: Javier M. Mellid <jmunhoz@igalia.com>