[LIBCLOUD-873] Updated ProfitBricks Compute Driver (REST api v3)
Refactoring the compute  driver to remove support for the EOL SOAP api and modifying api calls to use the REST api (v3).

Removing XML test fixtures and replacing them with JSON fixtures.

Providing wrappers for all new functionality provided in v3 of the Cloud API.

Addressing an issue in libcloud/common/base.py where a http status code of 202 would not be flagged as a successful request.