Create Simplified Chinese Version of README.rst
Hello, I translated the README.rst to Simplified Chinese using Google Translate and my own understanding in Simplified Chinese language. I hope this can help people who are not fluent in English but are fluent in Chinese to be able to understand the project easier and faster. There are some words that I didn't convert to Chinese either because they are special names, or specific links which does not have a translated version ( for example at line 68: LICENSE_ an NOTICE_ ). The file name is README-zh.rst, but I don't mind if its changed, its just for easier initial identification.

Thank you.