Reintroduce S3 multipart upload support with signature v4
All S3 regions now support signature V4. Rather than try to implement multipart
uploads with both V2 and V4, all region drivers have been converted to use
the latter. All multipart changes are based on code which existed in v1.5.0.

Some minor adjustments were made to the use of requests made throughout
the BaseS3StorageDriver class. The practice of constructing request paths with
parameters included needed to be removed. It causes issues during the
signature calculation. The solution is to use the params argument instead.
In addition, a fix was necessary in the signature calculation to not compute a
payload hash in the event of file uploads. It seems this was already taken into
consideration but failed to account for the scenario where an iterator is
passed to the data argument.

Closes #1005