Fix improper getresponse() uses
Closes #1133

During the migration from httplib to requests, a lot of driver code did
not change because they use the high-level libcloud API which hid the
change from httplib to requests with the HttpLibResponseProxy adapter.
However, in some cases, driver code needs to access the actual response,
and does this by calling getresponse().

While this used to return an httlib.HTTPResponse instance, this now
returns a requests.Response instance. All code covered by tests was
fixed, but OVH and Brightbox common code is not fully covered. I don't
have a subscription to either service and can't test the change, but the
resulting code can only be more correct since it fixes the
`getresponse() result comes from httplib` assumption.