Ensure RawResponse downloads don't consume RAM
The fix is general but the cause is very specific: since libcloud 2.0,
`download_object()` in the S3 and GCS drivers first places the item in
RAM, then writes it to a file. The reason it broke is that accessing
`body` in a requests Response consumes it entirely and loads it in RAM.
To fix this, I moved `body` to a property: it is still accessibly, but
loaded only if requested.

Closes #1132

Signed-off-by: Tomaz Muraus <tomaz@tomaz.me>