Reuse TCP connections when uploading files
It's easy to break connection reuse when using the requests API: just
use `stream=True` and never read the response. The connection used to
make the request will never be reused, and will be dropped when the
urllib3's connection pool is full.

It turns out uploading objects using the S3 API goes through
`prepared_request`, which incorrectly sets `stream` to the value of
`raw`, `True` in our case. And since we don't read the response data,
the connection are never reused, and each upload requires its own
connection.

This is particularly wasteful when uploading many small objects, which
can easily happen with JSON or Parquet files generated by Apache Spark,
where setting up the connection takes significant time compared to
uploading a few bytes.

Setting `stream=stream` in the `prepared_request` method matches the
code in the `request` method and fixes the bug.