OpenStack v2 create_volume can create bootable vol
This PR adds `ex_image_ref` to create_volume for the
OpenStack_2_NodeDriver. This allows the user to create
bootable volumes based on machine images. When the UUID
of an image is passed as ImageRef it will create a bootable
volume from which later an instance can be launched. This is
helpful because for older versions of OpenStack it is not
possible to create an instance with a block device mapping
where the newly created volume has to be put in a availability
zone that is not the default.

More info about bootable volumes at https://docs.openstack.org/api-ref/block-storage/v2/index.html?expanded=create-volume-detail#volumes-volumes

> To create a bootable volume, include the UUID of the image from which
> you want to create the volume in the imageRef attribute in the request
> body.