Label support for GCP volumes
There was no support in the GCP provider for querying/setting labels on volumes - this adds that!

In order to set the labels for a disk, the user is required to pass in the last labelFingerprint, which must be fetched first. This is done when building the StorageVolume object from a query, pull out the relevant fields (labels, labelFingerprint) and stuff them in the `extra` dictionary.

In addition, a new method has been added, `ex_set_volume_labels` which sets the labels for a given volume.