Make id lower for consistent uuid generation
When a node is created, the resource groups name in the id string is lower case. In subsequent calls, like in list_nodes(), for some reason Azure returns the resource groups name in upper case. This results in a different uuid from when create_node is called. This also breaks wait_until_running(). Forcing lower case in _to_node() seems to resolve this issue.