test/test_http.py: stop HTTP server (thread) properly
The HttpLayerTestCase test class creates a HTTPServer instance and runs
it in a separate thread.

After running all the test-cases, we attempt to close the server and the
thread by killing the thread. Unfortunately, the code that does that is
unreachable because the tear-down member function is called tearDownCls
intead of tearDownClass.

Moreover, there is no threading.Thread.kill function. This was undetected
because the code was unreachable.

The proper way to clean things up is to:

1. Stop the HTTP server using HTTPServer.shutdown()
2. Join the thread using threading.Thread.join()