test/test_connection.py: do not clear http_proxy unconditionally
The tearDownClass function from BaseConnectionClassTestCase used to
clear the http_proxy environment variable unconditionally. This can have
unexpected side effects since subsequent tests run with no http_proxy
set even if it was inherited from the environment when pytest was
started.

Also, the http_proxy variable is already restored to its saved value (or
cleared if there was no value to save) in the tearDown function after
each test is executed, so there seems to be no reason to attempt to
clear it again.

Likewise, the libcloud.common.base.ALLOW_PATH_DOUBLE_SLASHES variable is
already reset to False after each test is executed so there is no reason
to set it to False again after all tests from the TestCase have
executed.