test/test_connection.py: allow tests to run if https_proxy is set
We have code in setUp & tearDown functions to make tests run properly if
the http_proxy variable is already present in the environment when
running tests.

The code backs up the http_proxy variable, clears it, and restores it
after the tests have run.

However, we would also need to handle the HTTPS counterpart, the
https_proxy variable. Running tests with this variable set currently
fails with:

      self.assertIsNone(conn.proxy_scheme)
E       AssertionError: 'http' is not None

libcloud/test/test_connection.py:127: AssertionError

This patch adds the same logic that we have for HTTP, but for the HTTPS
counterpart.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>