test: alleviate race conditions in test_retry_*
The retry tests (test_retry_*) are inherently racy because they assume mock side effects
will always finish before the timeout. If for some reason the execution is a bit slow,
this can lead to the timeout beating the mocks, and the tests will fail. Adjust these tests
to alleviate this issue.