Replace "skip-duplicate-action" GHA jobs pre-check with native GHA "concurrency" functionality.
Concurrency functionality which allows us to cancel redundant builds is
now available natively as part of GHA functionality so we can get rid of
skip duplicate actions action and simplify the workflow.

This way we can also get rid of write permission which was needed by
that action to be able to cancel in progress GHA job runs.