add vpc and elastic ips to ecs node response
As per documentation [1], the InnerIpAddress field is only relevant
when the node network type is set to classic. If the node is within
a VPC, the field used is PrivateIpAddress within VpcAttributes.

In addition, elastic ips in field EipAddress are also public ips.
They differ from other public ips in that they are static and stay
the same once allocated.

Fields EipAddress and VpcAttributes are already returned in the response
as part of the "extra" field of the node object. This is however not
consistent with the other drivers as these fields are also public and
private ip addresses, respectively.

[1] https://www.alibabacloud.com/help/en/ecs/developer-reference/api-describeinstances