[azure arm] also delete the VM OS disk if it is a managed disk
VMs on Azure can be launched with managed or unamanaged disks.
Unmanaged disks are typically launched from specific VHDs.
Managed disks are managed by Azure and they do not have a VHD
associated that we can remove (e.g., ubuntu image offered by Azure).

Deleting a VM on the Azure Web console allows you to choose whether
you want to delete the OS disk as well, independently of this is
managed or unmanaged. The original code would allow that only for
unmanaged disks via the 'ex_destroy_vhd' parameter but, as said,
this is incomplete as VMs can be launched from Azure managed disks.

This change proposes to introduce a new parameter 'ex_destroy_os_disk'
to delete the VM OS disk independently of their type. Ideally, this
parameter should default to True as (i) that is the Azure's default
option and (ii) it would align with 'ex_destroy_vhd'. However, this
might pose backward compatibility issues if libcloud starts deleting
managed OS disks by default as well. I am not sure what the libcloud
maintainers' take is in this regard but this would be my proposal.